# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClassPulse AI — Real-time classroom intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated teacher
      responses:
        "200":
          description: Current teacher profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update the current teacher's profile (name, subject, avatarUrl)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeBody"
      responses:
        "200":
          description: Updated teacher profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
...
  /auth/select-teacher:
    post:
      operationId: selectTeacher
      tags: [auth]
      summary: Issue an auth token by teacher id (no password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectTeacherBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "404":
          description: Teacher not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/admin-login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin password login — returns an admin-role token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Admin login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid admin password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/select-student:
    post:
      operationId: selectStudent
      tags: [auth]
      summary: Issue a student-role token by student id (no password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectStudentBody"
      responses:
        "200":
          description: Student login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentLoginResponse"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
          enum: [teacher, admin, student]
        teacher:
          $ref: "#/components/schemas/Teacher"
      required: [token, teacher]
    SelectTeacherBody:
      type: object
      properties:
        teacherId:
          # Mongo Wise ID (string). Server still accepts legacy ints for back-compat.
          type: string
      required: [teacherId]
    SelectStudentBody:
      type: object
      properties:
        studentId:
          type: string
      required: [studentId]
    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
          description: Ignored in preview deployments — kept for backward compatibility.
      required: []
    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
          enum: [admin]
        admin:
          type: object
          properties:
            name: { type: string }
            email: { type: string }
          required: [name, email]
      required: [token, role, admin]
    StudentLoginResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
          enum: [student]
        student:
          $ref: "#/components/schemas/StudentOption"
      required: [token, role, student]
    StudentOption:
      type: object
      properties:
        id:
          # Mongo Wise student ID (24-char hex string)
          type: string
        name:
          type: string
        subject:
          type: string
          nullable: true
        grade:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
      required: [id, name]
    UpdateMeBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        subject:
          type: string
          minLength: 1
          maxLength: 120
        avatarUrl:
          type: string
          nullable: true
...
    Teacher:
      type: object
      properties:
        id:
          # Mongo Wise teacher ID (string) for Wise-sourced teachers;
          # numeric id (still serialized as string) for legacy Supabase rows.
          type: string
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        avatarUrl:
          type: string
          nullable: true
        totalClasses:
          type: integer
        avgScore:
          type: number
        role:
          type: string
          enum: [teacher, admin, student]
        grade:
          type: string
          nullable: true
        primaryTeacherId:
          type: integer
          nullable: true
      required: [id, name, email, subject, totalClasses, avgScore]
...